``` refactor(examples): add package comments and improve error handling
- Add descriptive package comments to all example files
- Add proper error checking for AddFrame operations in JPEG2000 examples
- Handle errors gracefully with return statements instead of panicking
- Remove unused variables and improve code safety

fix(jpeg): remove unnecessary error returns from convertToPixels

- Change convertToPixels method to return only pixel data without error
- Remove redundant error handling in Decode function since conversion doesn't fail
- Clean up code by removing unnecessary error variable declarations

refactor(jpeg): rename ExtendedCodec to Codec for consistency

- Rename ExtendedCodec struct to Codec to match naming conventions
- Update all method receivers and references accordingly
- Maintain same functionality while improving naming consistency

refactor(jpeg2000): simplify parser read methods and error handling

- Modify read methods to return only error instead of (n, error) when length is known
- Simplify readComponentIndex to return only the index value
- Remove unnecessary error returns from readTileDataWithLength
- Update callers to handle simplified return values

style(jpeg): fix code alignment and formatting issues

- Adjust spacing in struct field alignment for better readability
- Fix comment indentation and spacing around operators
- Standardize comment formatting throughout the decoder
```

+++ jpeg2000/htj2k/uvlc_debug_test.go
@@ -86,7 +86,7 @@ func TestUVLCEncodeDecode(t *testing.T) {
 				if !ok {
 					t.Fatalf("Failed to read suffix")
 				}
-				decoded := uint32(3 + suffix)
+				decoded := 3 + suffix
 				t.Logf("Decoded (prefix=001): suffix=%d, u=%d", suffix, decoded)
 				if decoded != tt.u {
 					t.Errorf("Mismatch: got %d, want %d", decoded, tt.u)
@@ -106,13 +106,13 @@ func TestUVLCEncodeDecode(t *testing.T) {
 				if !ok {
 					t.Fatalf("Failed to read extension")
 				}
-				decoded := uint32(5 + suffix + 4*ext)
+				decoded := 5 + suffix + 4*ext
 				t.Logf("Decoded (prefix=000, ext): suffix=%d, ext=%d, u=%d", suffix, ext, decoded)
 				if decoded != tt.u {
 					t.Errorf("Mismatch: got %d, want %d", decoded, tt.u)
 				}
 			} else {
-				decoded := uint32(5 + suffix)
+				decoded := 5 + suffix
 				t.Logf("Decoded (prefix=000): suffix=%d, u=%d", suffix, decoded)
 				if decoded != tt.u {
 					t.Errorf("Mismatch: got %d, want %d", decoded, tt.u)
